Add Loader, add category and fix migrations, do all to be able to use the application

+++ FinancialAccounting/Models/NoDBModels/CreateAccounts.cs
@@ -5,7 +5,7 @@
         public string AccountName { get; set; }
         public int AccountTypeId { get; set; }
         public bool isActiv { get; set; }
-        public long Balance { get; set; }
+        public double Balance { get; set; }
         public double? Interest { get; set; }
         public int? InterestInterval { get; set; }
         public DateTime? InterestAccrualDate { get; set; }
@@ -16,7 +16,7 @@
         public string AccountName { get; set; }
         public int AccountTypeId { get; set; }
         public bool isActiv { get; set; }
-        public long Balance { get; set; }
+        public double Balance { get; set; }
         public double? Interest { get; set; }
         public int? InterestInterval { get; set; }
         public DateTime? InterestAccrualDate { get; set; }
